Upgrade cryptokit to cryptokit-1.5
Repackage the cryptokit-1.5 source with a single change to allow building
without the OASIS build system, findlib, and their dependencies.

The Makefile from cryptokit-1.3 was copied to cryptokit-1.5/src and patched to
build the blowfish code.

The main SKS Makefile was changed to support building cryptokit-1.5 in the src
subdirectory